cast to double to avoid integer division

+++ src/object/character.cpp
@@ -732,9 +732,9 @@ void Character::died( vector< Object * > & objects ){
 			
 			int x = (int) getX() + part.x;
 			int y = (int) getY() + part.y;
-			double dx = (Util::rnd(11) - 5) / 4.2;
-			double dy = (Util::rnd(10) + 4) / 3.5;
-                        double dz = (Util::rnd(11) - 5) / 5;
+			double dx = (double)(Util::rnd(11) - 5) / 4.2;
+			double dy = (double)(Util::rnd(10) + 4) / 3.5;
+                        double dz = (double)(Util::rnd(11) - 5) / 5.0;
 			objects.push_back( new Gib( x, y, (int) getZ(), dx, dy, dz, part.image ) );
 		}
 		if ( squish_sound != NULL ){
